+++ build1/_deps/fmt-subbuild/fmt-populate-prefix/tmp/fmt-populate-mkdirs.cmake
@@ -6,22 +6,22 @@ cmake_minimum_required(VERSION ${CMAKE_VERSION}) # this file comes with cmake
 # If CMAKE_DISABLE_SOURCE_CHANGES is set to true and the source directory is an
 # existing directory in our source tree, calling file(MAKE_DIRECTORY) on it
 # would cause a fatal error, even though it would be a no-op.
-if(NOT EXISTS "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/fmt-src")
-  file(MAKE_DIRECTORY "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/fmt-src")
+if(NOT EXISTS "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/fmt-src")
+  file(MAKE_DIRECTORY "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/fmt-src")
 endif()
 file(MAKE_DIRECTORY
-  "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/fmt-build"
-  "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/fmt-subbuild/fmt-populate-prefix"
-  "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/fmt-subbuild/fmt-populate-prefix/tmp"
-  "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/fmt-subbuild/fmt-populate-prefix/src/fmt-populate-stamp"
-  "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/fmt-subbuild/fmt-populate-prefix/src"
-  "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/fmt-subbuild/fmt-populate-prefix/src/fmt-populate-stamp"
+  "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/fmt-build"
+  "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/fmt-subbuild/fmt-populate-prefix"
+  "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/fmt-subbuild/fmt-populate-prefix/tmp"
+  "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/fmt-subbuild/fmt-populate-prefix/src/fmt-populate-stamp"
+  "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/fmt-subbuild/fmt-populate-prefix/src"
+  "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/fmt-subbuild/fmt-populate-prefix/src/fmt-populate-stamp"
 )
 
 set(configSubDirs Debug)
 foreach(subDir IN LISTS configSubDirs)
-    file(MAKE_DIRECTORY "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/fmt-subbuild/fmt-populate-prefix/src/fmt-populate-stamp/${subDir}")
+    file(MAKE_DIRECTORY "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/fmt-subbuild/fmt-populate-prefix/src/fmt-populate-stamp/${subDir}")
 endforeach()
 if(cfgdir)
-  file(MAKE_DIRECTORY "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/fmt-subbuild/fmt-populate-prefix/src/fmt-populate-stamp${cfgdir}") # cfgdir has leading slash
+  file(MAKE_DIRECTORY "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/fmt-subbuild/fmt-populate-prefix/src/fmt-populate-stamp${cfgdir}") # cfgdir has leading slash
 endif()