+++ build2/_deps/asp2-build/cmake_install.cmake
@@ -1,8 +1,8 @@
-# Install script for directory: C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/asp2-src
+# Install script for directory: C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/asp2-src
 
 # Set the install prefix
 if(NOT DEFINED CMAKE_INSTALL_PREFIX)
-  set(CMAKE_INSTALL_PREFIX "C:/Program Files (x86)/xdBot2")
+  set(CMAKE_INSTALL_PREFIX "C:/Program Files (x86)/geobot2")
 endif()
 string(REGEX REPLACE "/$" "" CMAKE_INSTALL_PREFIX "${CMAKE_INSTALL_PREFIX}")
 
@@ -38,16 +38,16 @@ if(NOT DEFINED CMAKE_OBJDUMP)
 endif()
 
 if(CMAKE_INSTALL_COMPONENT STREQUAL "Unspecified" OR NOT CMAKE_INSTALL_COMPONENT)
-  file(INSTALL DESTINATION "${CMAKE_INSTALL_PREFIX}/lib" TYPE STATIC_LIBRARY FILES "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/asp2-build/asp.lib")
+  file(INSTALL DESTINATION "${CMAKE_INSTALL_PREFIX}/lib" TYPE STATIC_LIBRARY FILES "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/asp2-build/asp.lib")
 endif()
 
 if(CMAKE_INSTALL_COMPONENT STREQUAL "Unspecified" OR NOT CMAKE_INSTALL_COMPONENT)
-  file(INSTALL DESTINATION "${CMAKE_INSTALL_PREFIX}/include" TYPE DIRECTORY FILES "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/asp2-src/include/")
+  file(INSTALL DESTINATION "${CMAKE_INSTALL_PREFIX}/include" TYPE DIRECTORY FILES "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/asp2-src/include/")
 endif()
 
 string(REPLACE ";" "\n" CMAKE_INSTALL_MANIFEST_CONTENT
        "${CMAKE_INSTALL_MANIFEST_FILES}")
 if(CMAKE_INSTALL_LOCAL_ONLY)
-  file(WRITE "C:/Users/Benno111/Documents/GitHub/xdBot/build/_deps/asp2-build/install_local_manifest.txt"
+  file(WRITE "C:/Users/Benno111/Documents/GitHub/geobot/build/_deps/asp2-build/install_local_manifest.txt"
      "${CMAKE_INSTALL_MANIFEST_CONTENT}")
 endif()