+++ build2/cmake_install.cmake
@@ -1,8 +1,8 @@
-# Install script for directory: C:/Users/Benno111/Documents/GitHub/xdBot
+# Install script for directory: C:/Users/Benno111/Documents/GitHub/geobot
 
 # Set the install prefix
 if(NOT DEFINED CMAKE_INSTALL_PREFIX)
-  set(CMAKE_INSTALL_PREFIX "C:/Program Files (x86)/xdBot2")
+  set(CMAKE_INSTALL_PREFIX "C:/Program Files (x86)/geobot2")
 endif()
 string(REGEX REPLACE "/$" "" CMAKE_INSTALL_PREFIX "${CMAKE_INSTALL_PREFIX}")
 
@@ -39,13 +39,13 @@ endif()
 
 if(NOT CMAKE_INSTALL_LOCAL_ONLY)
   # Include the install script for the subdirectory.
-  include("C:/Users/Benno111/Documents/GitHub/xdBot/build/geode/cmake_install.cmake")
+  include("C:/Users/Benno111/Documents/GitHub/geobot/build/geode/cmake_install.cmake")
 endif()
 
 string(REPLACE ";" "\n" CMAKE_INSTALL_MANIFEST_CONTENT
        "${CMAKE_INSTALL_MANIFEST_FILES}")
 if(CMAKE_INSTALL_LOCAL_ONLY)
-  file(WRITE "C:/Users/Benno111/Documents/GitHub/xdBot/build/install_local_manifest.txt"
+  file(WRITE "C:/Users/Benno111/Documents/GitHub/geobot/build/install_local_manifest.txt"
      "${CMAKE_INSTALL_MANIFEST_CONTENT}")
 endif()
 if(CMAKE_INSTALL_COMPONENT)
@@ -61,6 +61,6 @@ else()
 endif()
 
 if(NOT CMAKE_INSTALL_LOCAL_ONLY)
-  file(WRITE "C:/Users/Benno111/Documents/GitHub/xdBot/build/${CMAKE_INSTALL_MANIFEST}"
+  file(WRITE "C:/Users/Benno111/Documents/GitHub/geobot/build/${CMAKE_INSTALL_MANIFEST}"
      "${CMAKE_INSTALL_MANIFEST_CONTENT}")
 endif()