+++ src/gdr/gdr.hpp
@@ -14,7 +14,7 @@ cocos2d::CCPoint dataFromString(std::string dataString);
 
 std::vector<std::string> splitByChar(std::string str, char splitChar);
 
-const std::string xdBotVersion = "v2.5.0-alpha.1";
+const std::string geobotVersion = "v2.5.0-alpha.1";
 
 namespace gdr {
 
@@ -159,10 +159,10 @@ namespace gdr {
 				replay.framerate = replayJson["framerate"];
 
 			bool rotation = ver.find("beta.") == std::string::npos && ver.find("alpha.") == std::string::npos;
-			if (replay.botInfo.name == "xdBot" && ver == "v2.0.0") rotation = true;
+			if (replay.botInfo.name == "geobot" && ver == "v2.0.0") rotation = true;
 
 			// bool offset = false;
-			int offset = replay.botInfo.name == "xdBot" ? 1 : 0;
+			int offset = replay.botInfo.name == "geobot" ? 1 : 0;
 
 			if (offset == 1) {
 				if (ver.front() == 'v') ver = ver.substr(1);
@@ -227,7 +227,7 @@ namespace gdr {
 					frameFix.p2.rotate = false;
 
 				} else if (frameFixJson.contains("p1")) {
-					if (replay.botInfo.name != "xdBot") rotation = false;
+					if (replay.botInfo.name != "geobot") rotation = false;
 
 					if (frameFixJson["p1"].contains("x"))
 						frameFix.p1.pos.x = frameFixJson["p1"]["x"];