+++ src/macro.hpp
@@ -37,13 +37,13 @@ struct input : gdr::Input {
 
 struct Macro : gdr::Replay<Macro, input> {
 
-    Macro() : Replay("xdBot", xdBotVersion.c_str()) {}
+    Macro() : Replay("geobot", geobotVersion.c_str()) {}
 
 public:
 
     bool canChangeFPS = true;
     uintptr_t seed = 0;
-    bool xdBotMacro = true;
+    bool geobotMacro = true;
 
     static void recordAction(int frame, int button, bool player2, bool hold);