+++ src/renderer/renderer.cpp
@@ -603,8 +603,8 @@ void Renderer::stop(int frame) {
         if (pl->m_isPaused && audioMode == AudioMode::Record) {
             if (PauseLayer* layer = Global::getPauseLayer()) {
                 CCScene* scene = CCDirector::sharedDirector()->getRunningScene();
-                if (RecordLayer* xdbot = scene->getChildByType<RecordLayer>(0))
-                    xdbot->onClose(nullptr);
+                if (RecordLayer* geobot = scene->getChildByType<RecordLayer>(0))
+                    geobot->onClose(nullptr);
                 
                 layer->onResume(nullptr);
             }