+++ changelog.md
@@ -118,7 +118,7 @@
 * Swapped positions of sfx and music volume in render settings.
 * Changed discord server popup message.
 * Re-added .gdr save toggle.
-* Re-added xdBot menu button in mod settings.
+* Re-added geobot menu button in mod settings.
 * Moved Show Trajectory extra settings to its own UI.
 * Moved Auto Stop Playing setting to more settings.
 * Moved Lock Delta setting to more settings.
@@ -417,7 +417,7 @@
 * Added Show Trajectory.
 * Added Frame Offset setting.
 * Added Customizable Keybinds.
-* Plus every feature from xdBot v1 (except frame fixes).
+* Plus every feature from geobot v1 (except frame fixes).
 
 # v1.6.1
 
@@ -488,7 +488,7 @@
  * Fixed occasional crash on restart.
  * Fixed macros ignoring frame fixes when loaded.
  * Made it so speedhack sets to 1 when you play a macro to avoid confusion.
- * Made it so xdBot button always appears at the end screen if the setting is enabled.
+ * Made it so geobot button always appears at the end screen if the setting is enabled.
 
 # v1.4.5